executor: move file migration to separate module
Refactor file migration and migrated file cleaning into a separate
module. This is done in preparation for the implementation of overlay
migration support.
