+++ tests/unit/executor/test_part_handler.py
@@ -481,82 +481,3 @@ class TestHelpers:
     def test_remove_non_existent(self):
         # this should not raise and exception
         part_handler._remove(Path("not_here"))
-
-    def test_clean_shared_area(self, new_dir):
-        p1 = Part("p1", {"plugin": "dump", "source": "subdir1"})
-        Path("subdir1").mkdir()
-        Path("subdir1/foo.txt").write_text("content")
-
-        p2 = Part("p2", {"plugin": "dump", "source": "subdir2"})
-        Path("subdir2").mkdir()
-        Path("subdir2/foo.txt").write_text("content")
-        Path("subdir2/bar.txt").write_text("other content")
-
-        info = ProjectInfo(application_name="test", cache_dir=new_dir)
-
-        handler1 = PartHandler(
-            p1, part_info=PartInfo(info, part=p1), part_list=[p1, p2]
-        )
-        handler2 = PartHandler(
-            p2, part_info=PartInfo(info, part=p2), part_list=[p1, p2]
-        )
-
-        for step in [Step.PULL, Step.BUILD, Step.STAGE]:
-            handler1.run_action(Action("p1", step))
-            handler2.run_action(Action("p2", step))
-
-        part_states = part_handler._load_part_states(Step.STAGE, part_list=[p1, p2])
-
-        assert Path("stage/foo.txt").is_file()
-        assert Path("stage/bar.txt").is_file()
-
-        part_handler._clean_shared_area(
-            part_name="p1", shared_dir=p1.stage_dir, part_states=part_states
-        )
-
-        assert Path("stage/foo.txt").is_file()  # remains, it's shared with p2
-        assert Path("stage/bar.txt").is_file()
-
-        part_handler._clean_shared_area(
-            part_name="p2", shared_dir=p2.stage_dir, part_states=part_states
-        )
-
-        assert Path("stage/foo.txt").exists() is False
-        assert Path("stage/bar.txt").exists() is False
-
-    def test_clean_migrated_files(self, new_dir):
-        Path("subdir").mkdir()
-        Path("subdir/foo.txt").touch()
-        Path("subdir/bar").mkdir()
-
-        p1 = Part("p1", {"plugin": "dump", "source": "subdir"})
-        info = ProjectInfo(application_name="test", cache_dir=new_dir)
-        part_info = PartInfo(info, part=p1)
-        handler = PartHandler(p1, part_info=part_info, part_list=[p1])
-
-        handler.run_action(Action("p1", Step.PULL))
-        handler.run_action(Action("p1", Step.BUILD))
-        handler.run_action(Action("p1", Step.STAGE))
-
-        assert Path("stage/foo.txt").is_file()
-        assert Path("stage/bar").is_dir()
-
-        part_handler._clean_migrated_files({"foo.txt"}, {"bar"}, Path("stage"))
-
-        assert Path("stage/foo.txt").exists() is False
-        assert Path("stage/bar").exists() is False
-
-    def test_clean_migrated_files_missing(self, new_dir):
-        Path("subdir").mkdir()
-
-        p1 = Part("p1", {"plugin": "dump", "source": "subdir"})
-        info = ProjectInfo(application_name="test", cache_dir=new_dir)
-        part_info = PartInfo(info, part=p1)
-        handler = PartHandler(p1, part_info=part_info, part_list=[p1])
-
-        handler.run_action(Action("p1", Step.PULL))
-        handler.run_action(Action("p1", Step.BUILD))
-        handler.run_action(Action("p1", Step.STAGE))
-
-        # this shouldn't raise an exception
-        part_handler._clean_migrated_files({"foo.txt"}, {"bar"}, Path("stage"))
